fix the deduct from student and add photo to course

+++ app/Http/Controllers/Api/v1/User/FavouriteController.php
@@ -19,6 +19,9 @@ class FavouriteController extends Controller
     public function index()
     {
         $user = auth()->user();
+         if(!$user){
+             return $this->error_response('Unauthenticated', [], 401);
+        }
         $favorite = $user->favourites;
         return $this->success_response('Available favorite', $favorite);
     }
@@ -29,7 +32,12 @@ class FavouriteController extends Controller
             'product_id'=>'required|exists:products,id'
         ]);
 
-        $favorite = ProductFavourite::where('user_id',$request->user()->id)
+         $user = auth()->user();
+         if(!$user){
+             return $this->error_response('Unauthenticated', [], 401);
+        }
+
+        $favorite = ProductFavourite::where('user_id',$user->id)
             ->where('product_id',$request->product_id)->first();
         if($favorite){
             if ($favorite->delete()) {
@@ -39,7 +47,7 @@ class FavouriteController extends Controller
             }
         }
         $favorite = new ProductFavourite();
-        $favorite->user_id = $request->user()->id;
+        $favorite->user_id = $user->id;
         $favorite->product_id = $request->product_id;
         if ($favorite->save()) {
             return response(['message' => 'Changed','is_favorite'=>true], 200);
@@ -48,37 +56,5 @@ class FavouriteController extends Controller
         }
     }
 
-
-    public function indexProvider()
-    {
-        $user = auth()->user();
-        $favorite = $user->providerFavourites;
-        return $this->success_response('Available Provider Favourites', $favorite);
-    }
-
-    public function storeProvider(Request $request)
-    {
-        $this->validate($request,[
-            'provider_type_id'=>'required|exists:provider_types,id'
-        ]);
-
-        $favorite = ProviderFavourite::where('user_id',$request->user()->id)
-            ->where('provider_type_id',$request->provider_type_id)->first();
-        if($favorite){
-            if ($favorite->delete()) {
-                return response(['message' => 'Changed','is_favorite'=>false], 200);
-            }else{
-                return response(['errors' => ['Something wrong']], 403);
-            }
-        }
-        $favorite = new ProviderFavourite();
-        $favorite->user_id = $request->user()->id;
-        $favorite->provider_type_id = $request->provider_type_id;
-        if ($favorite->save()) {
-            return response(['message' => 'Changed','is_favorite'=>true], 200);
-        }else{
-            return response(['errors' => ['Something wrong']], 403);
-        }
-    }
 
 }
